requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyDrive==1.3.1
pyparsing==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
Telethon==1.38.1
uritemplate==4.1.1
urllib3==2.2.3
